Allows DVP token's holders to increase their position.
Also improves Vault security.

Now the users of a DVP must specify the amount to burn even when the position
expired; this is due to the fact that the PositionManager holds every
position of its users as one single position within the DVP and with that
a PositionManager user was able to redeem all the payoffs with just one token.