Preliminary work for keeping the average volatility for the next epoch.
This should allow the DVP based on tokens for which we don't have data about
the implied volatility at the oracle level to use the average one of the
previous epoch trades as replacement.

In order to protect from missing data from the oracle we still need to provide
an initial value for the implied volatility of those tokens.

This should also mitigate the issues raised by TRST-L-10.